remove old `Page1` and change `Page2` into `Page`.
