+++ ProtoDraw/ProtoDraw.csproj
@@ -26,7 +26,6 @@
     <None Remove="OtherPage.xaml" />
     <None Remove="SettingsPage.xaml" />
     <None Remove="Views\DirectNPage.xaml" />
-    <None Remove="Views\DirectNPage2.xaml" />
   </ItemGroup>
 
   <ItemGroup>
@@ -89,9 +88,6 @@
   <ItemGroup>
     <Page Update="DirectNPage.xaml">
       <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Update="Views\DirectNPage2.xaml">
-      <XamlRuntime>$(DefaultXamlRuntime)</XamlRuntime>
     </Page>
     <Page Update="Views\DirectNPage.xaml">
       <XamlRuntime>$(DefaultXamlRuntime)</XamlRuntime>
