+++ ProtoDraw/Views/MainWindow.xaml.cs
@@ -26,13 +26,9 @@ namespace DirectNXAML.Views
             var item = args.InvokedItemContainer as NavigationViewItem;
             if (args != null)
             {
-                if ((string)(item.Tag) == "DirectNPage1")
+                if ((string)(item.Tag) == "DirectNPage")
                 {
                     ContentFrame.Navigate(typeof(DirectNPage), null);
-                } else
-                if ((string)(item.Tag) == "DirectNPage2")
-                {
-                    ContentFrame.Navigate(typeof(DirectNPage2), null);
                 }
                 else
                 {
